refactoring changes from feedback
