+++ app/constants/network.js
@@ -5,3 +5,4 @@ export const RINKEBY = 'rinkeby';
 export const GOERLI = 'goerli';
 export const RPC = 'rpc';
 export const NO_RPC_BLOCK_EXPLORER = 'NO_BLOCK_EXPLORER';
+export const PRIVATENETWORK = 'PRIVATENETWORK';
